# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.UpdateUserDto:
    properties:
      notify_account_activity:
        type: boolean
      notify_license_expiry:
        type: boolean
      notify_system_announcements:
        type: boolean
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  dto.UpdateUserRoleDto:
    properties:
      role:
        enum:
        - User
        - Manager
        - Admin
        type: string
    required:
    - role
    type: object
  dto.UpdateUserStatusDto:
    properties:
      status:
        enum:
        - Unverified
        - Verified
        - Active
        - Blocked
        type: string
    required:
    - status
    type: object
  errors.ApiError:
    properties:
      details: {}
      message:
        type: string
      status:
        type: integer
      type:
        $ref: '#/definitions/errors.ErrorType'
    type: object
  errors.ErrorType:
    enum:
    - VALIDATION_ERROR
    - NOT_FOUND_ERROR
    - UNAUTHORIZED_ERROR
    - FORBIDDEN_ERROR
    - CONFLICT_ERROR
    - INTERNAL_ERROR
    - BAD_REQUEST_ERROR
    - RATE_LIMIT_ERROR
    type: string
    x-enum-varnames:
    - ValidationError
    - NotFoundError
    - UnauthorizedError
    - ForbiddenError
    - ConflictError
    - InternalError
    - BadRequestError
    - RateLimitError
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: A high-performance license management system built with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: License Management API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all users (Admin only)
      parameters:
      - description: Page index (default 1)
        in: query
        name: pageIndex
        type: integer
      - description: Page size (default 10, max 100)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user account (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/errors.ApiError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve a specific user's details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/errors.ApiError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user profile information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ApiError'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/errors.ApiError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Change a user's role (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRoleDto'
      produces:
      - application/json
      responses:
        "200":
          description: User role updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ApiError'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/errors.ApiError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Update user role
      tags:
      - Users
  /users/{id}/status:
    patch:
      consumes:
      - application/json
      description: Change a user's account status (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserStatusDto'
      produces:
      - application/json
      responses:
        "200":
          description: User status updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ApiError'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/errors.ApiError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Update user status
      tags:
      - Users
  /users/by-license/{key}:
    get:
      consumes:
      - application/json
      description: Find the user associated with a specific license key (Admin only)
      parameters:
      - description: License key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/errors.ApiError'
        "404":
          description: License or user not found
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Get user by license key
      tags:
      - Users
  /users/export:
    get:
      consumes:
      - application/json
      description: Download all users in CSV format (Admin only)
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema:
            type: file
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/errors.ApiError'
      security:
      - BearerAuth: []
      summary: Export all users as CSV
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: JWT token in the format "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"